chore: fix opening
