+++ projects/demo/src/modules/components/input-chip/examples/3/index.ts
@@ -1,24 +1,26 @@
+import {NgIf} from '@angular/common';
 import {Component} from '@angular/core';
 import {FormControl, ReactiveFormsModule} from '@angular/forms';
 import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
-import {TuiHint, TuiIcon, TuiTextfield} from '@taiga-ui/core';
-import {TuiInputChip, TuiTooltip} from '@taiga-ui/kit';
+import type {TuiBooleanHandler} from '@taiga-ui/cdk';
+import {TuiButton, TuiTextfield} from '@taiga-ui/core';
+import {TuiInputChip} from '@taiga-ui/kit';
 
 @Component({
     standalone: true,
-    imports: [
-        ReactiveFormsModule,
-        TuiHint,
-        TuiIcon,
-        TuiInputChip,
-        TuiTextfield,
-        TuiTooltip,
-    ],
+    imports: [NgIf, ReactiveFormsModule, TuiButton, TuiInputChip, TuiTextfield],
     templateUrl: './index.html',
     encapsulation,
     changeDetection,
 })
 export default class Example {
-    protected readonly control = new FormControl(['Keep', 'it', 'simple']);
+    protected readonly required = ['Required', 'Obligatory'];
+
+    protected readonly control = new FormControl(this.required.concat('Removable'), {
+        nonNullable: true,
+    });
+
+    protected readonly handler: TuiBooleanHandler<string> = (item) =>
+        this.required.includes(item);
 }
