+++ projects/demo/src/modules/components/input-chip/examples/4/index.html
@@ -1,24 +1,103 @@
+<tui-textfield multi>
+    <input
+        placeholder="Type for suggestions"
+        tuiInputChip
+        [(ngModel)]="arbitrary"
+    />
+    <tui-input-chip *tuiItem />
+    <ng-container *ngIf="items | tuiHideSelected | tuiFilterByInput as items">
+        <ng-template
+            *ngIf="items.length"
+            tuiTextfieldDropdown
+        >
+            <tui-data-list-wrapper [items]="items" />
+        </ng-template>
+    </ng-container>
+</tui-textfield>
+
+<tui-textfield
+    #textfield
+    multi
+    [content]="!textfield.focused() && pythons.length ? 'Selected ' + pythons.length + ' out of ' + items.length : ''"
+    [disabledItemHandler]="handler"
+>
+    <label tuiLabel>Select Pythons</label>
+    <input
+        tuiInputChip
+        [placeholder]="pythons.length ? '' : 'Type for suggestions'"
+        [(ngModel)]="pythons"
+    />
+    <ng-template
+        *ngIf="!textfield.focused()"
+        tuiItem
+    />
+    <ng-container *ngIf="items | tuiHideSelected | tuiFilterByInput as items">
+        <ng-template
+            *ngIf="items.length"
+            tuiTextfieldDropdown
+        >
+            <tui-data-list-wrapper [items]="items" />
+        </ng-template>
+    </ng-container>
+</tui-textfield>
+
 <tui-textfield
     multi
-    [disabledItemHandler]="handler"
-    [style.width.rem]="19"
-    [tuiTextfieldCleaner]="false"
+    tuiChevron
+>
+    <label tuiLabel>Multi Select</label>
+    <input
+        tuiInputChip
+        tuiSelectLike
+        [placeholder]="multi.length ? '' : 'Pick from the list'"
+        [(ngModel)]="multi"
+    />
+    <tui-data-list-wrapper
+        *tuiTextfieldDropdown
+        new
+        tuiMultiSelectGroup
+        [items]="items"
+    />
+</tui-textfield>
+
+<tui-textfield
+    multi
+    tuiChevron
+    [disabledItemHandler]="strings"
+    [stringify]="stringify"
 >
     <input
         tuiInputChip
-        [formControl]="control"
+        [placeholder]="objects.length ? '' : 'Picking objects'"
+        [(ngModel)]="objects"
     />
-    <button
-        *ngIf="control.value.length > 2"
-        appearance="icon"
-        iconStart="@tui.x"
-        size="xs"
-        tabindex="-1"
-        tuiIconButton
-        type="button"
-        (click)="control.setValue(this.required)"
-    >
-        Clear
-    </button>
     <tui-input-chip *tuiItem />
+    <tui-data-list *tuiTextfieldDropdown>
+        <tui-opt-group
+            label="Pythons"
+            tuiMultiSelectGroup
+        >
+            <button
+                *ngFor="let user of users | tuiFilterByInput"
+                new
+                tuiOption
+                [value]="user"
+            >
+                {{ user.name }}
+            </button>
+        </tui-opt-group>
+        <tui-opt-group
+            label="Collaborators"
+            tuiMultiSelectGroup
+        >
+            <button
+                *ngFor="let user of more | tuiFilterByInput"
+                new
+                tuiOption
+                [value]="user"
+            >
+                {{ user.name }}
+            </button>
+        </tui-opt-group>
+    </tui-data-list>
 </tui-textfield>
