+++ projects/demo/src/modules/components/input-chip/index.html
@@ -44,23 +44,10 @@
         </tui-doc-example>
 
         <tui-doc-example
-            id="customization"
-            heading="Customization"
-            [component]="3 | tuiComponent"
-            [content]="3 | tuiExample"
-            [description]="customization"
-        >
-            <ng-template #customization>
-                You can customize many aspects of the component, from standard textfield options like icons and tooltips
-                to changing the appearance of each individual chip.
-            </ng-template>
-        </tui-doc-example>
-
-        <tui-doc-example
             id="disabled-items"
             heading="Disabled items"
-            [component]="4 | tuiComponent"
-            [content]="4 | tuiExample"
+            [component]="3 | tuiComponent"
+            [content]="3 | tuiExample"
             [description]="disabled"
         >
             <ng-template #disabled>
@@ -76,8 +63,8 @@
         <tui-doc-example
             id="dropdown"
             heading="Dropdown"
-            [component]="5 | tuiComponent"
-            [content]="5 | tuiExample"
+            [component]="4 | tuiComponent"
+            [content]="4 | tuiExample"
             [description]="dropdown"
         >
             <ng-template #dropdown>
@@ -85,6 +72,19 @@
                 can act as multiple value input with suggestions or as multi-select. With
                 <code>disabledItemHandler</code>
                 you can force the component to only accept suggested values.
+            </ng-template>
+        </tui-doc-example>
+
+        <tui-doc-example
+            id="customization"
+            heading="Customization"
+            [component]="5 | tuiComponent"
+            [content]="5 | tuiExample"
+            [description]="customization"
+        >
+            <ng-template #customization>
+                You can customize many aspects of the component, from standard textfield options like icons and tooltips
+                to changing the appearance of each individual chip.
             </ng-template>
         </tui-doc-example>
